template stubs for home and login, auth checking pseudocode in views
