+++ ScatterPeas/scatterpeas/templates/base.jinja2
@@ -1,0 +1,3 @@
+<h1>Our Reminder System</h1>
+{% block body %}
+{% endblock %}
