+++ ScatterPeas/scatterpeas/templates/home.jinja2
@@ -1,4 +1,4 @@
-{% extends base.jinja2 %}
+{% extends "base.jinja2" %}
 {% block body %}
 <h2>Hello World</h2>
 {% endblock %}
